feat: open full-screen mode at specified monitor (#77)
* feat: allow user choose monitor in settings

This commit fixes the issue that user open full-screen mode while
specify the monitor to use.

## Details

`libgdx` doesn't provide an api allows us to open on specified monitor,
but provides an api allows us speicfy the window position before
opening. This commit's appoarch is first open the window in `boardless`
mode and move the window to the speicifed monitor. Then once the window
is created, entering full-screen mode.

> Why not `windowed` mode but `boardless` mode?
> `windowed` mode ships with a `title bar`, which would make the game
window' content moves up after entering full-screen, it looks like some
of the game's content is being cutted.

## Drop-in Compatebility

This commit handles the cases that `monitor` option is empty, cannot be
found to keep the drop-in compatebiliity. The only possible break is the
upstream beatoraja also provides `monitor` option one day, which may
have different definition that conflicts the one from this pr.

* refactor: simply the steps for choosing the `DisplayMode`

* Fix case where both monitors report the same GLFW name

---------

Co-authored-by: Sarah A <sera@madotsuki.net>