Reuse FunctionBar's ability to make leaderboard bars selectable
This commit removes the `IRPlayerBar` class, and reuse `FunctionBar`
instead. The former one brings the functionality of selectable directly.
In this commit it's still an empty slot because we haven't made #111
stable.

However, to use the `FunctionBar` we need to extract it as a standalone
class, which leads to a bunch of changes.