IR service's leaderboard doesn't show player's name
When querying leaderboard score from IR service, it returns an empty
string as the name for current player's score. This is an intentional
design from upstream. Therefore, in this commit, name is manually set to
the player's own score.

Also this commit removes the `main` reference in `LR2IRConnection`,
since we delayed the name set step to "build leaderboard's bars" phase.